+++ zwave-js/zwave-js.js
@@ -1207,7 +1207,7 @@ module.exports = function (RED) {
 					Send(undefined, 'CONTROLLER_RESET_COMPLETE', undefined, send);
 					break;
 
-				case 'healNode':
+				case 'rebuildNodeRoutes':
 					NodeCheck(Params[0]);
 					ReturnNode.id = Params[0];
 					Send(ReturnNode, 'NODE_HEAL_STARTED', undefined, send);
@@ -1217,7 +1217,9 @@ module.exports = function (RED) {
 						text: 'Node Heal Started: ' + Params[0]
 					});
 					UI.Status('Node Heal Started: ' + Params[0]);
-					const HealResponse = await Driver.controller.healNode(Params[0]);
+					const HealResponse = await Driver.controller.rebuildNodeRoutes(
+						Params[0]
+					);
 					if (HealResponse) {
 						SetFlowNodeStatus({
 							fill: 'green',
@@ -1242,8 +1244,8 @@ module.exports = function (RED) {
 					RestoreReadyStatus();
 					break;
 
-				case 'beginHealingNetwork':
-					await Driver.controller.beginHealingNetwork();
+				case 'beginRebuildingRoutes':
+					await Driver.controller.beginRebuildingRoutes();
 					Send(undefined, 'NETWORK_HEAL_STARTED', undefined, send);
 					SetFlowNodeStatus({
 						fill: 'yellow',
@@ -1253,8 +1255,8 @@ module.exports = function (RED) {
 					UI.Status('Network Heal Started.');
 					break;
 
-				case 'stopHealingNetwork':
-					await Driver.controller.stopHealingNetwork();
+				case 'stopRebuildingRoutes':
+					await Driver.controller.stopRebuildingRoutes();
 					Send(undefined, 'NETWORK_HEAL_STOPPED', undefined, send);
 					SetFlowNodeStatus({
 						fill: 'blue',
@@ -1879,16 +1881,16 @@ module.exports = function (RED) {
 			let IsolatedNodeId;
 
 			if (Node !== undefined) {
-				PL.node = Node.id;
+				PL.node = Node.nodeId || Node.id;
 				IsolatedNodeId = Node.targetFlowNode || undefined;
 			}
 
 			if (Node !== undefined) {
-				const N = getNodeInfoForPayload(Node.id, 'name');
+				const N = getNodeInfoForPayload(Node.nodeId || Node.id, 'name');
 				if (N !== undefined) {
 					PL.nodeName = N;
 				}
-				const L = getNodeInfoForPayload(Node.id, 'location');
+				const L = getNodeInfoForPayload(Node.nodeId || Node.id, 'location');
 				if (L !== undefined) {
 					PL.nodeLocation = L;
 				}
@@ -1901,7 +1903,7 @@ module.exports = function (RED) {
 
 			let _Subject = '';
 			if (Node !== undefined) {
-				_Subject = '[Node: ' + Node.id + '] [' + Subject + ']';
+				_Subject = '[Node: ' + Node.nodeId || Node.id + '] [' + Subject + ']';
 			} else {
 				_Subject = '[' + Subject + ']';
 			}
@@ -1935,9 +1937,12 @@ module.exports = function (RED) {
 			];
 
 			if (TimestampSubjects.includes(Subject)) {
-				Driver.controller.nodes.get(Node.id).ZWNR_lastSeen = PL.timestamp;
-				Driver.controller.nodes.get(Node.id).ZWNR_lastEvent = PL.event;
-				Driver.controller.nodes.get(Node.id).ZWNR_lastObject = PL.object;
+				Driver.controller.nodes.get(Node.nodeId || Node.id).ZWNR_lastSeen =
+					PL.timestamp;
+				Driver.controller.nodes.get(Node.nodeId || Node.id).ZWNR_lastEvent =
+					PL.event;
+				Driver.controller.nodes.get(Node.nodeId || Node.id).ZWNR_lastObject =
+					PL.object;
 			}
 
 			if (AllowedSubjectsForDNs.includes(Subject) && SendDNs) {